Replaced BrandingValueResolver with RootContentResolver.
The reason for this is that the `BrandingValueResolver` would get the Root content node, then manually map it to the `Branding` property.
Whereas we can leverage Ditto to do this automatically.  If we return the Root content node from the value-resolver, then Ditto will know that an `IPublishedContent` is being returned
and will attempt to map it to the target type, (in this case the `Branding` object).

We could have kept the `BrandingValueResolver` class name and returned this...

```
return ContentResolver.Instance.GetRootContent().As<Branding>();
```

But I thought if we changed it to only return the Root content node, then it could potentially be reused.